NameDescription namespace fixes: models, query, controllers, views, routes, tests, fixtures
Also minor path fixes throughout helper, integration and model tests
