+++ app/models/description.rb
@@ -3,7 +3,7 @@
 #
 #  = Base Class for Models with Authored Descriptions
 #
-#  This class provides the common functionality between NameDescription and
+#  This class provides the common functionality between Name::Description and
 #  Location::Description.
 #
 #  == Class Methods
