+++ app/views/comments/_object.html.erb
@@ -14,7 +14,7 @@
     end
     tag.div str, class: "name-section m-3"
 
-  when "NameDescription"
+  when "Name::Description"
     render "names/name_description", object: object, review: false
 
   when "Observation"
