+++ app/views/users/show.html.erb
@@ -12,8 +12,8 @@
     [ :location_descriptions_editors, "location_descriptions_index_by_editor_path"  ],
     [ :locations,                     "locations_locations_by_user_path"      ],
     [ :locations_versions,            "locations_locations_by_editor_path"    ],
-    [ :name_descriptions_authors,     "name_descriptions_by_author_path"      ],
-    [ :name_descriptions_editors,     "name_descriptions_by_editor_path"      ],
+    [ :name_descriptions_authors,     "name_descriptions_index_by_author_path"      ],
+    [ :name_descriptions_editors,     "name_descriptions_index_by_editor_path"      ],
     [ :names,                         "names_names_by_user_path"              ],
     [ :names_versions,                'names_names_by_editor_path'            ],
     [ :observations,                  "observations_observations_by_user_path"],
