+++ script/refresh_sitemap
@@ -45,35 +45,35 @@ CONTROLLER_FILE = "#{::Rails.root}/app/controllers/NAME_controller.rb"
 OBJECT_TYPES = [
   ["observations",
    "Observation",
-   "observations/show_observation/ID"],
+   "observations/ID"],
   ["names",
    "Name",
-   "names/show_name/ID"],
+   "names/ID"],
   ["locations",
    "Location",
-   "locations/show_location/ID"],
+   "locations/ID"],
   ["species_lists",
    "SpeciesList",
-   "species_lists/show_species_list/ID"],
+   "species_lists/ID"],
   ["projects",
    "Project",
-   "projects/show_project/ID"],
+   "projects/ID"],
   ["glossary_terms",
    "GlossaryTerm",
-   "glossary/show_glossary_term/ID"],
+   "glossary/ID"],
   ["herbaria",
    "Herbarium",
-   "herbaria/show_herbarium/ID"],
+   "herbaria/ID"],
   ["name_descriptions",
-   "NameDescription",
-   "names/show_name_description/ID"],
+   "Name::Description",
+   "names/descriptions/ID"],
   ["location_descriptions",
    "Location::Description",
-   "locations/show_location_description/ID"]
+   "locations/descriptions/ID"]
 ].freeze
 
 STATIC_PAGES = [
-  ["home page",    "rss_log/list_rss_logs", "hourly"],
+  ["home page",    "observations",          "hourly"],
   ["introduction", "info/intro",            "weekly"],
   ["how to use",   "info/how_to_use",       "weekly"],
   ["how to help",  "info/how_to_help",      "weekly"]
