+++ test/helpers/name_helper_test.rb
@@ -2,7 +2,7 @@
 
 require "test_helper"
 
-# test the helpers for ObserverController
+# test the helpers for NamesController
 class NameHelperTest < ActionView::TestCase
   # Prove that the Query's used under "Observations of" (in the
   # show_name_info area) return the correct Observations
