+++ test/integration/capybara_lurker_test.rb
@@ -121,7 +121,7 @@ class CapybarLurkerTest < IntegrationTestCase
       #  Observation itself, naming, comment.
       # (plus a link to it is also in table of names for mobile)
       assert(
-        assert_selector("#content a[href^='/users/show_user/#{owner.id}']",
+        assert_selector("#content a[href^='/users/#{owner.id}']",
                         minimum: 4)
       )
 
@@ -144,21 +144,21 @@ class CapybarLurkerTest < IntegrationTestCase
                    page.title, "Wrong page")
 
       # (Make sure observation is shown somewhere.)
-      assert(has_selector?("a[href^='/#{obs.id}']"),
+      assert(has_selector?("a[href^='observations/#{obs.id}']"),
              "Missing a link to Observation")
     end # back at Observation
 
     # Check out Name
     go_back_after do
       # (Should be at least two links to show the Name.)
-      assert(assert_selector("#content a[href^='/names/show_name/#{name.id}']",
+      assert(assert_selector("#content a[href^='/names/#{name.id}']",
                              minimum: 2))
 
       click_link("About #{name.text_name}")
       # (Make sure the page contains create_name_description.)
       assert(
         assert_selector(
-          "#content a[href^='/names/create_name_description/#{name.id}']"
+          "#content a[href^='/names/#{name.id}/descriptions/new']"
         )
       )
     end # back at Observation
