+++ test/integration/curator_test.rb
@@ -8,14 +8,14 @@ class CuratorTest < IntegrationTestCase
     obs = observations(:minimal_unknown_obs)
     login!("mary", "testpassword", true)
     get("/#{obs.id}")
-    assert_template("observations/show_observation")
+    assert_template("observations/show")
     click(label: :create_herbarium_record.t)
-    assert_template("herbarium_records/create_herbarium_record")
+    assert_template("herbarium_records/new")
     open_form do |form|
       form.submit("Add")
     end
-    assert_template("observations/show_observation")
-    assert_match(%r{href="/observations/edit_observation/#{obs.id}},
+    assert_template("observations/show")
+    assert_match(%r{href="/observations/#{obs.id}/edit"},
                  response.body)
   end
 
@@ -24,29 +24,29 @@ class CuratorTest < IntegrationTestCase
     obs = observations(:detailed_unknown_obs)
     rec = obs.herbarium_records.find { |r| r.can_edit?(mary) }
     get("/#{obs.id}")
-    click(href: "/herbarium_record/edit_herbarium_record/#{rec.id}")
-    assert_template("herbarium_records/edit_herbarium_record")
+    click(href: "/herbarium_records/#{rec.id}/edit")
+    assert_template("herbarium_records/edit")
     open_form do |form|
       form.change("herbarium_name", "This Should Cause It to Reload Form")
       form.submit("Save")
     end
-    assert_template("herbarium_records/edit_herbarium_record")
+    assert_template("herbarium_records/edit")
     push_page
     open_form do |form|
       form.change("herbarium_name", rec.herbarium.name)
       form.submit("Save")
     end
-    assert_template("observations/show_observation")
-    assert_match(%r{href="/observations/edit_observation/#{obs.id}},
+    assert_template("observations/show")
+    assert_match(%r{href="/observations/#{obs.id}/edit"},
                  response.body)
     go_back
     click(label: "Cancel (Show Observation)")
-    assert_template("observations/show_observation")
-    assert_match(%r{href="/observations/edit_observation/#{obs.id}},
+    assert_template("observations/show")
+    assert_match(%r{href="/observations/#{obs.id}/edit"},
                  response.body)
-    click(href: "/herbarium_record/remove_observation/#{rec.id}")
-    assert_template("observations/show_observation")
-    assert_match(%r{href="/observations/edit_observation/#{obs.id}},
+    click(href: "/herbarium_records/remove_observation/#{rec.id}")
+    assert_template("observations/show")
+    assert_match(%r{href="/observations/#{obs.id}/edit"},
                  response.body)
     assert_not(obs.reload.herbarium_records.include?(rec))
   end
@@ -56,26 +56,26 @@ class CuratorTest < IntegrationTestCase
     obs = observations(:detailed_unknown_obs)
     rec = obs.herbarium_records.find { |r| r.can_edit?(mary) }
     get("/#{obs.id}")
-    click(href: "/herbarium_record/show_herbarium_record/#{rec.id}")
-    assert_template("herbarium_records/show_herbarium_record")
+    click(href: "/herbarium_records/#{rec.id}")
+    assert_template("herbarium_records/show")
     click(label: "Edit Herbarium Record")
-    assert_template("herbarium_records/edit_herbarium_record")
+    assert_template("herbarium_records/edit")
     open_form do |form|
       form.change("herbarium_name", "This Should Cause It to Reload Form")
       form.submit("Save")
     end
-    assert_template("herbarium_records/edit_herbarium_record")
+    assert_template("herbarium_records/edit")
     push_page
     click(label: "Cancel (Show Herbarium Record)")
-    assert_template("herbarium_records/show_herbarium_record")
+    assert_template("herbarium_records/show")
     go_back
     open_form do |form|
       form.change("herbarium_name", rec.herbarium.name)
       form.submit("Save")
     end
-    assert_template("herbarium_records/show_herbarium_record")
+    assert_template("herbarium_records/show")
     click(label: "Destroy Herbarium Record")
-    assert_template("herbarium_records/list_herbarium_records")
+    assert_template("herbarium_records/index")
     assert_not(obs.reload.herbarium_records.include?(rec))
   end
 
@@ -83,36 +83,36 @@ class CuratorTest < IntegrationTestCase
     login!("mary", "testpassword", true)
     obs = observations(:detailed_unknown_obs)
     rec = obs.herbarium_records.find { |r| r.can_edit?(mary) }
-    get("/herbarium/show_herbarium/#{rec.herbarium.id}")
-    click(href: /herbarium_index/)
-    assert_template("herbarium_records/list_herbarium_records")
-    click(href: "/herbarium_record/edit_herbarium_record/#{rec.id}")
-    assert_template("herbarium_records/edit_herbarium_record")
+    get("/herbaria/#{rec.herbarium.id}")
+    click(href: /herbaria/)
+    assert_template("herbarium_records/index")
+    click(href: "/herbarium_records/#{rec.id}/edit")
+    assert_template("herbarium_records/edit")
     open_form do |form|
       form.change("herbarium_name", "This Should Cause It to Reload Form")
       form.submit("Save")
     end
-    assert_template("herbarium_records/edit_herbarium_record")
+    assert_template("herbarium_records/edit")
     push_page
     click(label: "Back to Herbarium Record Index")
-    assert_template("herbarium_records/list_herbarium_records")
+    assert_template("herbarium_records/index")
     go_back
     open_form do |form|
       form.change("herbarium_name", rec.herbarium.name)
       form.submit("Save")
     end
-    assert_template("herbarium_records/list_herbarium_records")
-    click(href: "/herbarium_record/destroy_herbarium_record/#{rec.id}")
-    assert_template("herbarium_records/list_herbarium_records")
+    assert_template("herbarium_records/index")
+    click(href: "/herbarium_records/#{rec.id}") # FIXME select text DESTROY
+    assert_template("herbarium_records/index")
     assert_not(obs.reload.herbarium_records.include?(rec))
   end
 
   def test_herbarium_index_from_create_herbarium_record
     login!("mary", "testpassword", true)
-    get("/herbarium_record/create_herbarium_record/" +
+    get("/herbarium_records/new/" +
         observations(:minimal_unknown_obs).id.to_s)
     click(label: :herbarium_index.t)
-    assert_template("herbaria/list_herbaria")
+    assert_template("herbaria/index")
   end
 
   def test_single_herbarium_search
@@ -122,7 +122,7 @@ class CuratorTest < IntegrationTestCase
       form.select("type", :HERBARIA.l)
       form.submit("Search")
     end
-    assert_template("herbaria/show_herbarium")
+    assert_template("herbaria/show")
   end
 
   def test_multiple_herbarium_search
@@ -132,7 +132,7 @@ class CuratorTest < IntegrationTestCase
       form.select("type", :HERBARIA.l)
       form.submit("Search")
     end
-    assert_template("herbaria/list_herbaria")
+    assert_template("herbaria/index")
   end
 
   def test_herbarium_record_search
@@ -142,7 +142,7 @@ class CuratorTest < IntegrationTestCase
       form.select("type", :HERBARIUM_RECORDS.l)
       form.submit("Search")
     end
-    assert_template("herbarium_records/list_herbarium_records")
+    assert_template("herbarium_records/index")
   end
 
   def test_herbarium_change_code
@@ -151,7 +151,7 @@ class CuratorTest < IntegrationTestCase
     assert_not_equal(new_code, herbarium.code)
     curator = herbarium.curators[0]
     login!(curator.login, "testpassword", true)
-    get("/herbarium/edit_herbarium?id=#{herbarium.id}")
+    get("/herbaria/#{herbarium.id}/edit")
     open_form do |form|
       form.assert_value("code", herbarium.code)
       form.change("code", new_code)
@@ -159,14 +159,14 @@ class CuratorTest < IntegrationTestCase
     end
     herbarium = Herbarium.find(herbarium.id)
     assert_equal(new_code, herbarium.code)
-    assert_template("herbaria/show_herbarium")
+    assert_template("herbaria/show")
   end
 
   def test_herbarium_create
     user = users(:mary)
     assert_equal([], user.curated_herbaria)
     login!(user.login, "testpassword", true)
-    get("/herbarium/create_herbarium")
+    get("/herbaria/new")
     open_form do |form|
       form.assert_value("herbarium_name", "")
       form.assert_value("code", "")
@@ -181,6 +181,6 @@ class CuratorTest < IntegrationTestCase
     end
     user = User.find(user.id)
     assert_not_empty(user.curated_herbaria)
-    assert_template("herbaria/show_herbarium")
+    assert_template("herbaria/show")
   end
 end
