+++ test/integration/random_test.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+# TODO: NIMMO  Use filter_test selector patterns for links, if ambiguous here
 
 require "test_helper"
 
@@ -21,7 +22,7 @@ class RandomTest < IntegrationTestCase
 
   test "the homepage" do
     get("/")
-    assert_template("rss_logs/list_rss_logs")
+    assert_template("rss_logs/index")
     assert(/account/i, response.body)
   end
 
@@ -32,19 +33,19 @@ class RandomTest < IntegrationTestCase
     assert_template("info/how_to_help")
     assert_no_link_exists("/account/login")
     assert_link_exists("/account/logout_user")
-    assert_link_exists("/users/show_user/#{rolf.id}")
+    assert_link_exists("/users/#{rolf.id}")
 
     click(label: "Logout")
     assert_template("account/logout_user")
     assert_link_exists("/account/login")
     assert_no_link_exists("/account/logout_user")
-    assert_no_link_exists("/users/show_user/#{rolf.id}")
+    assert_no_link_exists("/users/#{rolf.id}")
 
     click(label: "How To Help")
     assert_template("info/how_to_help")
     assert_link_exists("/account/login")
     assert_no_link_exists("/account/logout_user")
-    assert_no_link_exists("/users/show_user/#{rolf.id}")
+    assert_no_link_exists("/users/#{rolf.id}")
   end
 
   test "sessions" do
