+++ test/integration/student_test.rb
@@ -2,7 +2,8 @@
 
 require "test_helper"
 
-# TODO: NIMMO the routing of name description urls has totally changed, fix
+# FIXME: NIMMO the routing of name description urls has totally changed, fix
+# FIXME: NIMMO  Use filter_test selector patterns for links, if ambiguous here
 # Test typical sessions of university student who is writing descriptions.
 class StudentTest < IntegrationTestCase
   # -----------------------------------
@@ -45,10 +46,10 @@ class StudentTest < IntegrationTestCase
       end
       assert_no_match(/#{gen_desc}/, response.body)
       assert_select("a[href*=create_name_description]", 1)
-      click(href: /show_name_description/)
+      click(href: /descriptions/)
       assert_select("a[href*=edit_name_description]")
       assert_select("a[href*=destroy_name_description]")
-      click(href: /edit_name_description/)
+      click(href: /edit/)
       open_form do |form|
         form.assert_value("source_type", "project")
         form.assert_value("source_name", project.title)
@@ -66,7 +67,7 @@ class StudentTest < IntegrationTestCase
   module CreatorDsl
     # Navigate to show name (no descriptions) and create draft.
     def create_draft(name, gen_desc, project)
-      assert_nil(NameDescription.find_by_gen_desc(gen_desc))
+      assert_nil(Name::Description.find_by_gen_desc(gen_desc))
       get("/")
       click(label: "Names", in: :left_panel)
       click(label: name.text_name)
@@ -94,12 +95,12 @@ class StudentTest < IntegrationTestCase
 
       # Make sure it shows up on main show_name page and can edit it.
       get(url)
-      assert_select("a[href*=edit_name_description]", 1)
-      assert_select("a[href*=destroy_name_description]", 1)
+      assert_select("a[href*=edit]", 1)
+      assert_select("a[href*=descriptions]", 1)
 
       # Now give it some text to make sure it *can* (but doesn't) actually get
       # displayed (content, that is) on main show_name page.
-      click(href: /edit_name_description/)
+      click(href: /edit/)
       open_form do |form|
         form.assert_value("source_type", :project)
         form.assert_value("source_name", project.title)
@@ -113,7 +114,7 @@ class StudentTest < IntegrationTestCase
         form.submit
       end
       assert_flash_success
-      assert_not_nil(NameDescription.find_by_gen_desc(gen_desc))
+      assert_not_nil(Name::Description.find_by_gen_desc(gen_desc))
       url
     end
   end
@@ -123,8 +124,8 @@ class StudentTest < IntegrationTestCase
     def check_another_student(url)
       get(url)
       click(href: /show_name_description/)
-      assert_select("a[href*=edit_name_description]", 0)
-      assert_select("a[href*=destroy_name_description]", 0)
+      assert_select("a[href*=edit]", 0)
+      assert_select("a[href*=descriptions]", 0)
     end
   end
 
@@ -132,8 +133,8 @@ class StudentTest < IntegrationTestCase
     # Knows it exists but can't even view it.
     def check_another_user(url)
       get(url)
-      assert_select("a[href*=show_name_description]", 1)
-      click(href: /show_name_description/)
+      assert_select("a[href*=descriptions]", 1)
+      click(href: /descriptions/)
       assert_flash_error
       assert_nil(assigns(:description))
     end
