+++ test/models/abstract_model_test.rb
@@ -125,7 +125,7 @@ class AbstractModelTest < UnitTestCase
     user_attrs     = user.attributes.dup
 
     num_past_names = Name.versioned_class.count
-    num_past_name_descs = NameDescription.versioned_class.count
+    num_past_name_descs = Name::Description.versioned_class.count
     num_past_locations = Location.versioned_class.count
     num_past_loc_descs = Location::Description.versioned_class.count
 
@@ -145,7 +145,7 @@ class AbstractModelTest < UnitTestCase
     end
 
     assert_equal(num_past_names + 0, Name.versioned_class.count)
-    assert_equal(num_past_name_descs + 0, NameDescription.versioned_class.count)
+    assert_equal(num_past_name_descs + 0, Name::Description.versioned_class.count)
     assert_equal(num_past_locations + 0, Location.versioned_class.count)
     assert_equal(num_past_loc_descs + 0,
                  Location::Description.versioned_class.count)
