+++ test/test_helper.rb
@@ -108,8 +108,10 @@ module ActiveSupport
     #
     # Note: You'll currently still have to declare fixtures explicitly
     # in integration tests -- they do not yet inherit this setting
-    self.set_fixture_class locations: Location
+    self.set_fixture_class locations: Location # I added. necessary? - AN
     self.set_fixture_class 'locations/descriptions' => Location::Description
+    self.set_fixture_class names: Name # I added. necessary? - AN
+    self.set_fixture_class 'names/descriptions' => Name::Description
     fixtures :all
 
     # Add more helper methods to be used by all tests here...
